+++ CHANGELOG.md
@@ -152,8 +152,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Labels for CSV sources
 - util.cli CMD's correcly set the description of subparsers instead of their
   help, they also accept the `CLI_FORMATTER_CLASS` property.
-- CSV source now has `entry_point` decoration
-- JSON source now has `entry_point` decoration
+- CSV source now has `entrypoint` decoration
+- JSON source now has `entrypoint` decoration
 - Strict flag in df.memory is now on by default
 - Dynamically created scikit models get config args correctly
 - Renamed `DNNClassifierModelContext` first init arg from `config` to `features`
@@ -182,7 +182,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - MemoryOperationImplementationNetwork instantiates OperationImplementations
   using their `withconfig()` method.
-- MemorySource now decorated with `entry_point`
+- MemorySource now decorated with `entrypoint`
 - MemorySource takes arguments correctly via `config_set` and `config_get`
 - skel modules have `long_description_content_type` set to "text/markdown"
 - Base Orchestrator `__aenter__` and `__aexit__` methods were moved to the